backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
apscheduler